requirements.txt
torch==2.3.1
torchvision
transformers
pandas
numpy
scikit-learn
nltk
Pillow
librosa
Flask
python-dotenv
nltk
tensorflow
keras
# re # uncomment if your python version does not have re built-in
